[FromRISO][Android] added base source code from RISO

+++ Android/3rdparty/pdfviewer/src/com/radaee/pdf/BMDatabase.java
@@ -7,16 +7,16 @@ class for PDF Book-mark database.
 
 public class BMDatabase
 {
-	private static native long openAndCreate( String db_path );
-	private static native void close( long db );
-	private static native long recOpen( long db, String look_path );
-	private static native void recClose( long rec );
-	private static native int recGetCount( long rec );
-	private static native String recItemGetName( long rec, int index );
-	private static native int recItemGetPage( long rec, int index );
-	private static native boolean recItemRemove( long rec, int index );
-	private static native boolean recItemInsert( long rec, String name, int pageno );
-	private long m_hand = 0;
+	private static native int openAndCreate( String db_path );
+	private static native void close( int db );
+	private static native int recOpen( int db, String look_path );
+	private static native void recClose( int rec );
+	private static native int recGetCount( int rec );
+	private static native String recItemGetName( int rec, int index );
+	private static native int recItemGetPage( int rec, int index );
+	private static native boolean recItemRemove( int rec, int index );
+	private static native boolean recItemInsert( int rec, String name, int pageno );
+	private int m_hand = 0;
 	/**
 	 * open exist database file or create it if not exist. (Database for BookMark)
 	 * @param db_path
@@ -40,7 +40,7 @@ public class BMDatabase
 	 * @param look_path PDF path to lookup.
 	 * @return handle value of RecordSet.
 	 */
-	public long RecOpen( String look_path )
+	public int RecOpen( String look_path )
 	{
 		return recOpen( m_hand, look_path );
 	}
@@ -48,7 +48,7 @@ public class BMDatabase
 	 * Close a RecordSet handle
 	 * @param rec handle value of RecordSet. obtained by RecOpen
 	 */
-	public void RecClose( long rec )
+	public void RecClose( int rec )
 	{
 		recClose( rec );
 	}
@@ -57,7 +57,7 @@ public class BMDatabase
 	 * @param rec RecordSet Handle, obtained by RecOpen
 	 * @return records count
 	 */
-	public int RecGetCount( long rec )
+	public int RecGetCount( int rec )
 	{
 		return recGetCount( rec );
 	}
@@ -67,7 +67,7 @@ public class BMDatabase
 	 * @param index 0 based index value, range:[0, RecGetCount()-1]
 	 * @return name of item.
 	 */
-	public String RecItemGetName( long rec, int index )
+	public String RecItemGetName( int rec, int index )
 	{
 		return recItemGetName( rec, index );
 	}
@@ -77,7 +77,7 @@ public class BMDatabase
 	 * @param index 0 based index value, range:[0, RecGetCount()-1]
 	 * @return 0 based page NO.
 	 */
-	public int RecItemGetPage( long rec, int index )
+	public int RecItemGetPage( int rec, int index )
 	{
 		return recItemGetPage( rec, index );
 	}
@@ -87,7 +87,7 @@ public class BMDatabase
 	 * @param index 0 based index value, range:[0, RecGetCount()-1]
 	 * @return true or false.
 	 */
-	public boolean RecItemRemove( long rec, int index )
+	public boolean RecItemRemove( int rec, int index )
 	{
 		return recItemRemove( rec, index );
 	}
@@ -98,7 +98,7 @@ public class BMDatabase
 	 * @param pageno 0 based page no recorded.
 	 * @return true or false
 	 */
-	public boolean RecItemInsert( long rec, String name, int pageno )
+	public boolean RecItemInsert( int rec, String name, int pageno )
 	{
 		return recItemInsert( rec, name, pageno );
 	}
